debug "NA" string in excel
1. Now NA is allowed in all fields except user accession.
2. user accession has to follow the user accession naming rule and must
be unique in a single excel file, or leave it blank if there is a
system accession. It also works if both system accession and user
accession present in a row, in this case, both accessions can be used
to establish relationships.
3. Now it complains and prints warning about invalid string (anything
not start with TRGT or USR) in relationship columns, and it won’t
attempt to connect them.