new version with upload and update modes
Now the script provides 2 different modes:
1, upload mode (default). All records with system accessions are
skipped; if the combination of user accession and user name exist in
our database for a given record, it will also be skipped. All other
records are uploaded to the database as new records.
2, update mode. records in excel without system accessions are skipped,
while records with system accession in the excel will cover the record
with same system accession in the database.