fetch matching user_accession/accession from database
Fetch matching user_accession/accession from database before
submit/update the records in the excel.
If only one of user_accession/accession is found in the excel file, the
script will try to find the matching other accession from the database,
and add it to the record. Then in the next step the redundant records
will be properly skipped.